Clean up pages / routes folder
