+++ application/account-management/WebApp/pages/(auth)/register.verify.tsx
@@ -1,0 +1,11 @@
+import { createFileRoute } from "@tanstack/react-router";
+import { HorizontalHeroLayout } from "@/shared/ui/layout/HorizontalHeroLayout";
+import { CompleteAccountRegistrationForm } from "@/shared/ui/auth/CompleteAccountRegistrationForm";
+
+export const Route = createFileRoute("/(auth)/register/verify")({
+  component: () => (
+    <HorizontalHeroLayout>
+      <CompleteAccountRegistrationForm />
+    </HorizontalHeroLayout>
+  )
+});
