+++ application/account-management/WebApp/pages/index.tsx
@@ -1,22 +1,22 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { CtaSection } from "./-landing/sections/CtaSection";
-import { FeatureSection } from "./-landing/sections/FeatureSection";
-import { FeatureSection2 } from "./-landing/sections/FeatureSection2";
-import { HeroSection } from "./-landing/sections/HeroSection";
-import { FeatureSection3 } from "./-landing/sections/FeatureSection3";
-import { TechnologySection } from "./-landing/sections/TechnologySection";
-import { TechnologySection2 } from "./-landing/sections/TechnologySection2";
-import { CommunitySection } from "./-landing/sections/CommunitySection";
-import { FeatureSection4 } from "./-landing/sections/FeatureSection4";
-import { CtaSection2 } from "./-landing/sections/CtaSection2";
-import { CtaSection3 } from "./-landing/sections/CtaSection3";
-import { FooterSection } from "./-landing/sections/FooterSection";
+import { CtaSection } from "../shared/ui/landingPage/sections/CtaSection";
+import { FeatureSection } from "../shared/ui/landingPage/sections/FeatureSection";
+import { FeatureSection2 } from "../shared/ui/landingPage/sections/FeatureSection2";
+import { HeroSection } from "../shared/ui/landingPage/sections/HeroSection";
+import { FeatureSection3 } from "../shared/ui/landingPage/sections/FeatureSection3";
+import { TechnologySection } from "../shared/ui/landingPage/sections/TechnologySection";
+import { TechnologySection2 } from "../shared/ui/landingPage/sections/TechnologySection2";
+import { CommunitySection } from "../shared/ui/landingPage/sections/CommunitySection";
+import { FeatureSection4 } from "../shared/ui/landingPage/sections/FeatureSection4";
+import { CtaSection2 } from "../shared/ui/landingPage/sections/CtaSection2";
+import { CtaSection3 } from "../shared/ui/landingPage/sections/CtaSection3";
+import { FooterSection } from "../shared/ui/landingPage/sections/FooterSection";
 
 export const Route = createFileRoute("/")({
-  component: Home
+  component: LandingPage
 });
 
-function Home() {
+function LandingPage() {
   return (
     <main className="flex w-full flex-col">
       <HeroSection />
