+++ application/account-management/WebApp/shared/ui/landingPage/sections/FooterSection.tsx
