+++ application/account-management/WebApp/shared/ui/landingPage/sections/HeroSection.tsx
