+++ application/account-management/WebApp/shared/ui/landingPage/sections/TechnologySection.tsx
